Create Password Generator using Python GUI programing
The Password Generator project is a simple application that can randomly generate passwords with a combination of letters, numbers, and special characters based on a length given by the user using Python's Tkinter module. The user can store the password along with their username in a database using Python's SQLite3 module. The program only accepts users whose usernames have not been previously stored. The GUI features a combobox to select the password length.